add basic webscraper
Added basic functionality to the webscraper. To install the required library, beautifulsoup4, on a python instance, use `pip install beautifulsoup4`.

Currently the scraper just prints the content(s) found inside of <nav> tags found on the what's@calu website.; however, inconsistency with how the academic calendar is formatted means it may be better to just create a pipeline for automatically adding events instead? So the school can have the events accurately reflect what's going on regardless of their html decisions.