Fix Progress
Statements used by the sql statements in this program successfully are interpreted by the db browser, but not by the sqllite3 library that python includes, and I have no idea why. This may have to do with how the connections to the database are resolved? I've found no concise guides on what's even happening here or how to fix it, just a lot of conjecture about what a safe string to pass to .execute() entails (you need to send a tuple of parameters and a string delimited by ?; that's what I'll try next after reformatting the function).