Webscraper Timeout Fix
I decided to edit the webscraper to only grab new events; which should solve our problems, as the .db included in the repo has every single file on the announcements page to date. This does mean we can't currently re-match events for new categories, but if that ends up being something we have time to add, I'll add it later.
The hardware handler now also properly isolates an event for suppress messages, which it hadn't done before. I was assuming if the screen wasn't cleared that it would loop it's display, but that likely isn't the case.